+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project tries to adhere to [Semantic Versioning](https://semver.org/spe
 - Single utility allowance for states with no BUA
 
 
+# 2021-06-03
+
+### Added
+- Non-citizen calculations
+
 # 2021-05-30
 
 ### Changed